+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - CUSTOMER INFORMATIONS : Indicate that Payment delay is in days
 - INVOICES DASHBOARD: Turnover is now calculated using both sales and assets
 - MANUF ORDER : fix missing form and grid view attributes for workshopStockLocation.
+- Stock Move Line : blocking message to delete an allocated stock move line
 
 ## Bug Fixes
 - Fix Timesheet Reminder Batch sendReminder method